lint-config: Add dedicated file for lint rules.
Add a dedicated file for lint rules. Keeping rules in package.json
is all well and good, but node uses that file for other specific
things, so it's probably a good idea to keep that file dedicated
to node things.

Also, keeping rules in the package.json file does not allow one
to add single or multi-line comments in an easy, non-confusing
way without screwing up node. As I'm sure we'd like to have some
documentation for our linting rules, an external JS file with
standard commenting capability is likely preferable.

Ref: UX-686